some fixes

+++ lab03/src/matrix_multiply.cpp
@@ -163,9 +163,12 @@ void matrixMultiplyOpenClImage(float *a, float *b, float *c, int m, int n, int k
     imageFormat.image_channel_order = CL_R;
     imageFormat.image_channel_data_type = CL_FLOAT;
 
-    auto aImage = clCreateImage2D(context, CL_MEM_READ_ONLY, &imageFormat, m, n, 0, nullptr, nullptr);
-    auto bImage = clCreateImage2D(context, CL_MEM_READ_ONLY, &imageFormat, n, k, 0, nullptr, nullptr);
-    auto cImage = clCreateImage2D(context, CL_MEM_READ_ONLY, &imageFormat, m, k, 0, nullptr, nullptr);
+    auto aImage = clCreateImage2D(context, CL_MEM_READ_ONLY, &imageFormat, static_cast<size_t>(m),
+                                  static_cast<size_t>(n), 0, nullptr, nullptr);
+    auto bImage = clCreateImage2D(context, CL_MEM_READ_ONLY, &imageFormat, static_cast<size_t>(n),
+                                  static_cast<size_t>(k), 0, nullptr, nullptr);
+    auto cImage = clCreateImage2D(context, CL_MEM_READ_ONLY, &imageFormat, static_cast<size_t>(m),
+                                  static_cast<size_t>(k), 0, nullptr, nullptr);
     size_t origin[] = {0, 0, 0};
     {
         size_t region[] = {static_cast<size_t>(m), static_cast<size_t>(n), 1};
